Phase_5/requirements.txt
flask
gunicorn
transformers
torch
Pillow
